KubeRay: Relocate files to enable controller extension with Kubebuilder
