+++ ray-operator/controllers/raycluster/common/pod_test.go
@@ -2,6 +2,7 @@ package common
 
 import (
 	"fmt"
+	"github.com/ray-project/kuberay/ray-operator/controllers/raycluster/utils"
 	"reflect"
 	"sort"
 	"strings"
@@ -10,9 +11,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 
-	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/api/raycluster/v1alpha1"
-	"github.com/ray-project/kuberay/ray-operator/controllers/utils"
-
+	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/apis/raycluster/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/pointer"
 )
