+++ ray-operator/controllers/raycluster/common/service_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/api/raycluster/v1alpha1"
+	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/apis/raycluster/v1alpha1"
 
 	"github.com/stretchr/testify/assert"
