+++ ray-operator/controllers/raycluster/raycluster_controller_fake_test.go
@@ -13,15 +13,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package raycluster
 
 import (
 	"context"
+	"github.com/ray-project/kuberay/ray-operator/controllers/raycluster/common"
 	"testing"
 
 	. "github.com/onsi/ginkgo"
-	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/api/raycluster/v1alpha1"
-	"github.com/ray-project/kuberay/ray-operator/controllers/common"
+	rayiov1alpha1 "github.com/ray-project/kuberay/ray-operator/apis/raycluster/v1alpha1"
 	"github.com/ray-project/kuberay/ray-operator/pkg/client/clientset/versioned/scheme"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/labels"
