+++ ray-operator/pkg/client/clientset/versioned/typed/raycluster/v1alpha1/fake/fake_raycluster.go
@@ -5,7 +5,7 @@ package fake
 import (
 	"context"
 
-	v1alpha1 "github.com/ray-project/kuberay/ray-operator/api/raycluster/v1alpha1"
+	v1alpha1 "github.com/ray-project/kuberay/ray-operator/apis/raycluster/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
